Move functions.py to utils package
